Move project to internals
