+++ alias/set.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"path"
 
+	"github.com/JPZ13/dpm/internal/project"
 	"github.com/JPZ13/dpm/parser"
-	"github.com/JPZ13/dpm/project"
 	"github.com/JPZ13/dpm/utils"
 )
