+++ alias/unset.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"path"
 
+	"github.com/JPZ13/dpm/internal/project"
 	"github.com/JPZ13/dpm/parser"
-	"github.com/JPZ13/dpm/project"
 	"github.com/JPZ13/dpm/utils"
 )
