+++ internal/project/deactivate.go
